move statusbar from /common/ to /browser/
