+++ src/vs/workbench/services/statusbar/browser/statusbar.ts
